Create Diary.cpp
Simple diary program.

We have four commands:
‫‪
start_day‬‬ ‫‪day/month/year‬‬: This command signals the start of a new day to the program. Everything that comes between a start_day command and the next command has to be stored as notes for that day.

show_day day/month/year: The program has to print all notes from the requested day.

show_the_longest_day: Print the date of the day with the most notes; a summary of the notes has to printed after the date.

show_the_best_day: Print the day with the most positive words in its notes.

How we summarize a day's notes: Print the logs in their entirety if its shorter than 20 characters, otherwise print the first 20 characters with "..." at the end.