Cinema schedule
A Top Down Design program.

The program receives the screening schedules for a number of cinemas and has to provide information from this schedule to the user. The user can request information using the four command below.

GET ALL MOVIES
Print the name of every single movie that has a screening in the CSV file in alphabetical order.

‫‪GET‬‬ ‫‪SCHEDULE‬‬ <MOVIE NAME>
Provide a weekly screening schedule for the movie in the input. The program has to output the schedule in a table in terminal and an HTML file.
We might have multiple screening of a movie in different cinemas at the same time.  In this scenario we have to choose a screening by these criteria in order:
1- Earliest start time. 2- Cheapest ticket price. 3- Smallest lexicographical cinema name.