Create A6_Mafia_Inheritance
This program simulates the game of Mafia using objected oriented programming and inheritance. It also handles errors by throwing exception classes and appropriate error messages.

We have the following roles in the game:
‫‪Joker: Doesn't belong in either mafia or citizen group, but if the game ends and everyone loses if they get voted out during daytime voting. 
Villager: Is a citizen and votes during days for a mafia suspect to get executed.
Detective: Is a citizen, and gets waken up every night after mafia polling finishes; they can choose an individual and ask the god (here the program plays the role of god) whether they're a citizen or a mafia member.
Doctor:‬‬ Is a citizen, and wakes up at nights after mafia voting. They choose one of the remaining players to save; if that player was chosen by mafia, they get saved by the doctor and survive the night.
‫‪Bulletproof: Is a citizen. They have an additional life for the night votings, which means that they can survive the first time that mafia votes to kill them. But they can't survive a second attempt.
‫‪Mafia: Every night they take a vote to eliminate one the remaining players.‬‬ 
‫‪Godfather: They're a member of mafia, their only distinction with a normal mafia member is that the program will identify them as a citizen if a detective asks about them at night‬‬.
‫‪Silencer‬‬: Is a mafia member. Every night they choose one of the players; that player can't participate in the polling that happens in the following day.

Commands:

create_game (<player_name>)*
At the start the game we receive the list of player roles and create a game.

assign_role <player_name> <role_name>
Assigns the role to the player.

‫‪start_game‬‬
Start the game and print all player names with their roles. It also has to print "Ready? Set! Go." after the players information.
At the start of each day the program has to print "Day" with the day number afterwards, then the day voting begins.

Day votes format: ‫<voter_name> <voteee_name>
Day voting ends when the program receives "end_vote". Then it has to report the death of the person who got voted out in the "<user_name> died" format.

The night begins after the daytime death announcement and we have to announce it in an identical format to the day. We're also required to print the name and the role of every player that has to wake up at night.
Subsequently the program receives all the night actions in the following format: <first_player> <second_player>
This action means that the first player has chosen the second player. For example if the first player is a doctor, the second player gets saved, but if they're a mafia member, the second player gets a vote added to their night tally.

The night ends with an "end_night" command. It has to report the results of the mafia voting like below and proceed to daytime.
"Mafia tried to kill <user_name>"

From day 2 onwards the program has to report the events of the previous in the following format:
"Day <day_num>
<user_name> was killed
Silenced (<user_name>)*
<user_name1> swapped characters with <user_name2>"

This game has an additional feature that allows the god to swap players' roles with the swap_charater command.
"‫‪swap_character‬‬ ‫<user_name1> <user_name2>
" swaps the role of player user_name1 with player user_name2. 

get_game_state
Prints the number of remaining mafia members and villagers (aka citizens).
"‫‪‫‪Mafia = <number_of_mafia>
Villager = <number_of_villager>"

The game continues until on the following conditions are satistfied:
1- All mafia members are killed.
2- The number of villagers becomes equal or smaller than the number of mafia members divided by two.
3- The joker dies in a daytime vote.